employee_directory/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-extensions==3.2.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.3.0
django-seed==0.3.1
Faker==15.1.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toposort==1.7
